codec_adapter: add local buffer
This patch adds local buffer to codec adapter. The aim of
this buffer is to damp any irregularity between pipeline
and codec. The difference in buffer sizes is one of the
examples here - thanks to additional local buffer we can
call codec processing several times during one period or
we can store several pipeline periods in local buffer and
process them once.

Signed-off-by: Marcin Rajwa <marcin.rajwa@linux.intel.com>